buildABoard end of night Commit.
We can now build a board
Currently working on saving what we make into a board.
After that I need to make it so we can load a board into the editor as well so we can change it and save again. Also make more checks for creation to make sure we don't create errors later in the actual game.